+++ packages/graphql-server-express/src/index.ts
@@ -1,6 +1,7 @@
 export {
   ExpressGraphQLOptionsFunction,
   ExpressHandler,
+  ExpressGraphiQLOptionsFunction,
   graphqlExpress,
   graphiqlExpress,
 } from './expressApollo';