+++ packages/graphql-server-koa/src/index.ts
@@ -1,1 +1,7 @@
-export { KoaGraphQLOptionsFunction, KoaHandler, graphqlKoa, graphiqlKoa } from './koaApollo';
+export {
+  KoaGraphQLOptionsFunction,
+  KoaHandler,
+  KoaGraphiQLOptionsFunction,
+  graphqlKoa,
+  graphiqlKoa,
+} from './koaApollo';