+++ packages/graphql-server-lambda/src/index.ts
@@ -1,6 +1,8 @@
 export {
   LambdaHandler,
   IHeaders,
+  LambdaGraphQLOptionsFunction,
+  LambdaGraphiQLOptionsFunction,
   graphqlLambda,
   graphiqlLambda,
 } from './lambdaApollo';